pyproject.toml
[tool.poetry]
name = "indexer-analyze"
version = "0.1.0"
description = ""
authors = ["Xevion <xevion@xevion.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.1.0"
structlog = "^25.3.0"
httpx = {extras = ["http2"], version = "^0.28.1"}
anyio = "^4.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
